Fix store for blocks
One problem that I introduced for depending too much on the compiler:

The store right now uses common/lru which uses `interface{}` as a key.
`GetBlock` was receiving a key as `int` and using that to access the
`lru`. However, `SetBlock` was using `int64` as a key and using that to
store the block in the `lru`. The store would never find it and there
would be no error because everything is `interface{}`.

And that, kids, it's why I don't like to use `interface{}` that much.
Two hours spent debugging something that a compiler could have told me
in two seconds.

This commit still doesn't fix all.

+++ src/node/core.go
@@ -563,6 +563,6 @@ func (c *Core) GetLastCommittedRoundEventsCount() int {
 	return c.poset.LastCommitedRoundEvents
 }
 
-func (c *Core) GetLastBlockIndex() int {
+func (c *Core) GetLastBlockIndex() int64 {
 	return c.poset.Store.LastBlockIndex()
 }
